% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chowTest.R
\name{ChowTest}
\alias{ChowTest}
\title{Chow Test for Structural Change}
\usage{
ChowTest(model, n1, n2 = NULL)
}
\arguments{
\item{model}{An 'lm' or 'glm' object.}

\item{n1}{The number of rows of the first part of the model until the suspected breakpoint.}

\item{n2}{An optional parameter of the number of rows of the second part of the model after the breakpoint.}
}
\value{
The Chow Test statistic and the associated p-value.
}
\description{
ChowTest() function applies Chow Structural Change test to 'lm' and 'glm' objects.
}
\examples{
y <- c(5*rnorm(30 , 100 , 5 ) + 60 , 2*rnorm(20 , 100 , 5)+20)
x <- 5*rnorm(50 , 100 , 5 ) + 20
model1 <- lm(y ~ x)
ChowTest(model = model1 , n1 = 30)
ChowTest(model = model1 , n1 = 30 , n2 = 20)

}
\references{
{Chow, G. C. (1960). Tests of equality between sets of coefficients in two linear regressions. Econometrica: Journal of the Econometric Society, 591-605.}
}
