% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runsTest.R
\name{RunsTest.character}
\alias{RunsTest.character}
\title{Wald–Wolfowitz (Runs) Test}
\usage{
\method{RunsTest}{character}(
  x,
  alpha = 0.05,
  alternative = c("two.sided", "greater", "less"),
  exact = NULL
)
}
\arguments{
\item{x}{A 'character' object.}

\item{alpha}{The level of significance. It is used only in small sample (exact) test , since for large sample tests (asymptotic) p-value will be provided.}

\item{alternative}{The tail of the test. "two.sided" , "greater" and "less" are available.}

\item{exact}{If TRUE exact critical values will be used  otherwise normal approximation will be returned. By default , if both n1 and n2 are smaller than or equal to 20 it is considered a small sample and the exact values will be used.}
}
\value{
Wald–Wolfowitz runs test results , the associated p-value or the critical values.
}
\description{
RunsTest Method for 'character' class.
}
