% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runsTest.R
\name{RunsTest}
\alias{RunsTest}
\title{Wald–Wolfowitz (Runs) Test}
\usage{
RunsTest(x, ...)
}
\arguments{
\item{x}{A 'numeric' , 'character 'or 'factor' object.}

\item{...}{Additional arguments.}
}
\value{
Wald–Wolfowitz runs test results , the associated p-value or the critical values.
}
\description{
Generic function for Wald–Wolfowitz runs test.
}
