% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runsTest.R
\name{RunsTest}
\alias{RunsTest}
\title{Wald–Wolfowitz (Runs) Test}
\usage{
RunsTest(x, ...)
}
\arguments{
\item{x}{A 'numeric' , 'character 'or 'factor' object.}

\item{...}{Additional arguments.}
}
\value{
Wald–Wolfowitz runs test results , the associated p-value or the critical values.
}
\description{
Generic function for Wald–Wolfowitz runs test.
}
\examples{
ch_vec <- sample(x = c("John" , "Jane" ) , size = 30 , replace = TRUE , prob = c(40 , 60))
RunsTest(ch_vec)

fac_vec <- as.factor(sample(x = c("John" , "Jane" ) , size = 30 , replace = TRUE , prob = c(40 , 60)))
RunsTest(fac_vec)

num_vec <- round(rnorm(n = 30 , mean = 100 , sd = 10 ))
RunsTest(num_vec)

}
